reorganized team data access interface to follow CA rules
