+++ src/main/java/use_case/team_entry/TeamEntryInteractor.java
@@ -3,7 +3,6 @@ package use_case.team_entry;
 import entity.Player;
 import entity.Team;
 import use_case.PlayerDataAccessInterface;
-import use_case.TeamDataAccessInterface;
 
 import java.util.ArrayList;
 import java.util.HashSet;
